Make svsjoin just fake a join
This'll remove a lot of duplication, and avoid missing checks in the
svsjoin path that were in the join path.
